requirements.txt
python-telegram-bot==20.7
openai==1.3.0
cachetools==5.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
logging==0.4.9.6
requests==2.31.0